refactored ODK CSV loading commands
