+++ va_explorer/va_data_management/management/commands/load_death_csv.py
@@ -2,8 +2,9 @@ import argparse
 import pandas as pd
 from django.core.management.base import BaseCommand, CommandError
 from va_explorer.va_data_management.models import ODKFormChoice, Death
-from va_explorer.va_data_management.utils.loading import normalize_dataframe_columns, normalize_string, normalize_value
-
+from va_explorer.va_data_management.utils.loading import (
+    normalize_dataframe_columns, normalize_string, load_odk_csv_to_model
+)
 
 class Command(BaseCommand):
     """Load a death CSV using the previously loaded ODK definition."""
@@ -32,65 +33,12 @@ class Command(BaseCommand):
         ]
         odk_map_columns = [c.strip() for c in odk_map_columns]
 
-        # Build odk_map (all normalized but case preserved)
-        all_choices = ODKFormChoice.objects.filter(form_name=norm_form_name)
-        odk_map = {}
-        for choice in all_choices:
-            field_name = normalize_string(choice.field_name)
-            value = normalize_value(choice.value)
-            if field_name not in odk_map:
-                odk_map[field_name] = {}
-            odk_map[field_name][value] = choice.label
-
-        def lookup_label(col, v):
-            """Map value to label with normalization and robust handling."""
-            v_norm = normalize_value(v)
-            if v_norm in odk_map.get(col, {}):
-                return odk_map[col][v_norm]
-            # Fallback
-            return v if pd.isnull(v) or str(v).lower() == "nan" else v_norm
-
-        # Perform label replacement
-        for col in odk_map_columns:
-            if col in df.columns and col in odk_map:
-                print(f"Mapping csv column '{col}' with odk_map keys {list(odk_map[col].keys())}")
-                print("BEFORE:", df[col].unique())
-                df[col] = df[col].map(lambda v, c=col: lookup_label(c, v))
-                print("AFTER:", df[col].unique())
-            elif col not in df.columns:
-                print(f"Column '{col}' not in DataFrame")
-            elif col not in odk_map:
-                print(f"Column '{col}' not in odk_map") 
-
-        # Remove duplicate columns after renaming
-        df = df.loc[:, ~df.columns.duplicated()]
-
-        # Only keep columns that are model fields
-        model_fields = {f.name for f in Death._meta.get_fields()}
-        df = df[[col for col in df.columns if col in model_fields]]
-
-        # Identify integer fields in the model
-        int_fields = [
-            f.name for f in Death._meta.get_fields()
-            if getattr(f, "get_internal_type", lambda: None)() in [
-                "IntegerField",
-                "BigIntegerField",
-                "SmallIntegerField",
-                "PositiveIntegerField",
-                "PositiveSmallIntegerField",
-            ]
-        ]
-
-        def nan_to_none_for_intfields(row, int_fields):
-            return {
-                k: (None if (k in int_fields and pd.isnull(v)) else v)
-                for k, v in row.items()
-            }
-
-        objects = [
-            Death(**nan_to_none_for_intfields(row, int_fields))
-            for row in df.to_dict(orient="records")
-        ]
+        objects = load_odk_csv_to_model(
+            df=df,
+            model=Death,
+            odk_choices_queryset=ODKFormChoice.objects.filter(form_name=norm_form_name),
+            odk_map_columns=odk_map_columns,
+            verbose=True  # Set to False to reduce output
+        )
         Death.objects.bulk_create(objects)
-
         self.stdout.write(f"Imported {len(objects)} records for death")
